+++ CHANGELOG.rst
@@ -1,6 +1,13 @@
 LATEST CHANGES
 ==============
 
+2022-12-09
+----------
+- Removed node parameter `general.resolution`, replaced by `general.grab_resolution`.
+- Added node parameter `general.pub_resolution` used to reduce node computation and message bandwidth.
+- Removed node parameters `video.img_downsample_factor` and `depth.depth_downsample_factor`. Use the new parameter `general.pub_resolution` instead.
+- Set default DEPTH MODE to NEURAL (4)
+
 2022-12-06
 ----------
 - Improved the `zed_camera.launch.py`